add caching validation

+++ modelio/models/state.py
@@ -14,8 +14,11 @@ import os
 import json
 from json import JSONDecodeError
 from typing import Any, Dict
+from modelio.exceptions.cache import CachePathError, NullCachePathError
 from modelio.typing import _PKey
 from modelio.models.base import BaseModel
+
+BP_ENV_VAR = "MODELIO_CACHE_BASE_PATH"
 
 
 class StateModel(BaseModel):
@@ -23,9 +26,9 @@ class StateModel(BaseModel):
     def __init__(self, _id: _PKey, name: str) -> None:
         super().__init__(_id)
         self.name = name
-        cache_base_path = os.environ.get('MODELIO_CACHE_BASE_PATH')
+        self.cache_base_path = os.environ.get(BP_ENV_VAR)
         # resolved path name includes model name and id
-        self._cpath = f'{cache_base_path}/{self.name}-{self.id}'
+        self._cpath = f'{self.cache_base_path}/{self.name}-{self.id}'
 
     @property
     def name(self) -> str:
@@ -38,7 +41,15 @@ class StateModel(BaseModel):
     def cache(self) -> None:
         """
         Serialize contents and save to cache as a json file
+
+        :raises NullCachePathError: _description_
+        :raises CachePathError: _description_
         """
+        if not self.cache_base_path:
+            raise NullCachePathError(
+                f"Caching basepath is unset. set the {BP_ENV_VAR} environment variable before using model caching")
+        if not os.path.exists(self.cache_base_path):
+            raise CachePathError(f"Caching base path {self.cache_base_path} does not exist")
         with open(self._cpath, 'w+') as json_file:
             cached_payload = self.serialize()
             json.dump(cached_payload, json_file)
